+++ example/lib/main.dart
@@ -1,3 +1,4 @@
+// ignore_for_file: avoid_print
 import 'package:flutter/material.dart';
 import 'package:string_contains/string_contains.dart';
 
@@ -28,13 +29,16 @@ class MyHomePage extends StatefulWidget {
 
 class _MyHomePageState extends State<MyHomePage> {
   final stringContainsBadWords = 'I am CEO, Bitch';
-  final stringContainsUrl = 'Please visit our website : https://betterx.io/';
+  final stringContainsUrl = 'Please visit our website : https://betterx.io/.';
   final stringContainsMultipleUrls =
       'Please visit our website : https://betterx.io/ and you can search for any thing on google : google.com, I love Dart : https://dart.dev/, you can practice on : https://dartpad.dev/. Flutter uses Dart, you can create your own app with Flutter : https://flutter.dev/, this is a good example of Flutter : https://flutter.dev/docs/get-started/.';
   final stringContainsEmail =
-      "BetterX.io : Let's build something Better, User-centered & beautiful together\n for more info contact us at : info@betterx.io. My Personal Email : ravi@workrush.co.uk\n A fake email : gedoye9446@roxoas.com";
-  final stringContainsPhoneNumber = 'My Phone Number : +91-0000000000';
+      "BetterX.io : Let's build something Better, User-centered & beautiful together\n for more info contact us at : info@betterx.io. My Personal Email : ravikumar2710999@gmail.com.\n A fake email : gedoye9446@roxoas.com.";
+  final stringContainsPhoneNumber = 'My Phone Number : +91-8433491441';
   final normalString = 'I love Dart and Flutter';
+  final stringContainsHashtags =
+      'I love #dart and #flutter. @flutter is #awesome';
+
   @override
   Widget build(BuildContext context) {
     return Scaffold(
@@ -118,13 +122,71 @@ class _MyHomePageState extends State<MyHomePage> {
                       ),
                 ),
                 const Divider(),
+                // String Contains Hashtags
+                Text(
+                  '$stringContainsHashtags contains hashtags : ${stringContainsHashtags.containsHashtag() == true ? 'Yes' : 'No'}\n get all hashtags:\n${stringContainsHashtags.getHashtags().join('\n')}',
+                  textAlign: TextAlign.center,
+                  style: Theme.of(context).textTheme.headline6?.copyWith(
+                        color: Theme.of(context).errorColor,
+                      ),
+                ),
+                const Divider(),
+                // String Contains maintains
+                Text(
+                  '$stringContainsHashtags maintains hashtags : ${stringContainsHashtags.containsMention() == true ? 'Yes' : 'No'}\n get all maintains:\n${stringContainsHashtags.getMentions().join('\n')}',
+                  textAlign: TextAlign.center,
+                  style: Theme.of(context).textTheme.headline6?.copyWith(
+                        color: Theme.of(context).errorColor,
+                      ),
+                ),
+                const Divider(),
                 CleanWidget(
+                  isSelecatable: true,
                   source:
-                      '$stringContainsBadWords\n $stringContainsUrl\n $stringContainsMultipleUrls\n $stringContainsEmail\n $stringContainsPhoneNumber\n $normalString',
+                      '$stringContainsBadWords\n$stringContainsUrl\n$stringContainsMultipleUrls\n$stringContainsEmail\n$stringContainsPhoneNumber\n$normalString',
                   obscuringCharacter: '\$',
                   style: Theme.of(context).textTheme.headline6?.copyWith(
                         color: Theme.of(context).primaryColorDark,
                       ),
+                ),
+                const Divider(),
+                Text(
+                  '$stringContainsBadWords\n$stringContainsUrl\n$stringContainsMultipleUrls\n$stringContainsEmail\n$stringContainsPhoneNumber\n$normalString\n$stringContainsHashtags',
+                ),
+                const Divider(),
+                StringContainsWidget(
+                  isSelecatable: true,
+                  source:
+                      '$stringContainsBadWords\n$stringContainsUrl\n$stringContainsMultipleUrls\n$stringContainsEmail\n$stringContainsPhoneNumber\n$normalString\n$stringContainsHashtags',
+                  onTap: (url) {
+                    print('element is a ${url.value} and type of ${url.type}');
+                    if (url.type == StringContainsElementType.words) {
+                      ScaffoldMessenger.of(context).showSnackBar(
+                        SnackBar(
+                          content: Text(
+                            '${url.value} is awesome!!',
+                          ),
+                          duration: const Duration(seconds: 2),
+                          dismissDirection: DismissDirection.horizontal,
+                        ),
+                      );
+                    }
+                  },
+                  highLightWords: const [
+                    'Flutter',
+                    'Dart',
+                  ],
+                  highlightWordsStyle: const TextStyle(
+                    color: Colors.red,
+                  ),
+                  types: const [
+                    StringContainsElementType.email,
+                    StringContainsElementType.url,
+                    StringContainsElementType.phoneNumber,
+                    StringContainsElementType.hashtag,
+                    StringContainsElementType.mention,
+                    StringContainsElementType.words,
+                  ],
                 ),
               ],
             ),