+++ lib/string_contains.dart
@@ -2,3 +2,7 @@ library string_contains;
 
 export 'src/string_contains.dart';
 export 'src/clean_widget.dart';
+export 'src/string_contains_widget.dart';
+
+export 'src/utils/constants.dart' show StringContainsElementType;
+export 'src/models/string_contains_element.dart' show StringContainsElement;