+++ CHANGELOG.md
@@ -5,3 +5,8 @@
 * Extension on String that provides a methods to check, if a string contains any bad-words/url/link/email/phone number/hashtag/mention and more others.
 * Added CleanWidget
 * Added methods to check if a string contains a hashtag, a mention.
+
+## 0.0.3
+* CleanWidget now supports a custom list of words to be removed.
+* StringContainsWidget added.
+* Updated methods to check if a string contains a hashtag, a mention.